+++ scripts/build.sh
@@ -68,7 +68,7 @@ cat > "$APP_BUNDLE/Contents/Info.plist" << EOF
 	<key>CFBundleIdentifier</key>
 	<string>${BUNDLE_ID}</string>
 	<key>CFBundleShortVersionString</key>
-	<string>1.2</string>
+	<string>1.3</string>
 	<key>CFBundleVersion</key>
 	<string>3</string>
 	<key>CFBundleExecutable</key>